backend/requirements.txt
openai>=1.37.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
requests>=2.32.0

